making graphics lib modular
