+++ graphics/run
@@ -1,7 +1,22 @@
 #!/usr/bin/env bash
-set -eux
-as -ggdb3 --32 -o entry.o entry.S
-gcc -c -ggdb3 -m16 -ffreestanding -fno-PIE -nostartfiles -nostdlib -o main.o -std=c99 main.c
-ld -m elf_i386 -o main.elf -T linker.ld entry.o main.o
-objcopy -O binary main.elf main.img
-qemu-system-x86_64 -drive file=main.img,format=raw
+
+# Implented the solution for compiling found here:
+# https://stackoverflow.com/questions/27669275/undefined-reference-to-in-os-kernel-linking
+
+nasm -f elf32 -Fdwarf -g boot.asm -o boot.o
+gcc -fno-PIC -g -c -m16 -ffreestanding -Os -Wall -fomit-frame-pointer main.c -o main.o
+ld -melf_i386 -nostartfiles -nostdlib -Tlinker.ld -o os.elf boot.o main.o
+
+# Convert os.elf to flat binary file os.bin
+objcopy -Obinary os.elf os.bin
+
+# Build 1.44MB disk image
+dd if=/dev/zero of=disk.img bs=1024 count=1440
+dd if=os.bin of=disk.img conv=notrunc
+
+# Split the boot sector from the complete os.bin file
+# These files may not be needed, generate them anyway
+dd if=os.bin of=boot.bin bs=512 count=1
+dd if=os.bin of=kernel.bin bs=512 seek=1
+
+qemu-system-i386 -fda disk.img
